🐛 [BE] 리퀘스트 타임 아웃 에러 잡은 걸지도..? 그 이외에도 셋팅 프롬프트 수정.

See merge request s09-final/S09P31A405!186

+++ backend/src/main/java/com/a405/gamept/util/dto/request/ChatGptForStreamRequestDto.java
@@ -7,7 +7,7 @@ import lombok.Builder;
 public record ChatGptForStreamRequestDto(
         String model,
         ChatGptMessage[] messages,
-        int temperature,
+        double temperature,
         int max_tokens,
         boolean stream
 ) {
